@@ -38,23 +38,23 @@ class StreamTest extends TestCase {
                 $this->elements = $elements;
             }
 
-            public function rewind() {
+            public function rewind(): void {
                 $this->position = 0;
             }
 
-            public function current() {
+            public function current(): mixed {
                 return $this->elements[$this->position];
             }
 
-            public function key() {
+            public function key(): mixed {
                 return $this->position;
             }
 
-            public function next() {
+            public function next(): void {
                 $this->position++;
             }
 
-            public function valid() {
+            public function valid(): bool {
                 return isset($this->elements[$this->position]);
             }
         };
